Patches by Todd Ingalls to compile on MacOS X.
git-svn-id: https://svn.code.sf.net/p/wxcl/svn/trunk@427 b8c11317-770e-0410-b8d1-8a1eaaf01d41

+++ wxc/src/ewxw/eljcolour.cpp
@@ -59,7 +59,7 @@ EWXWEXPORT(char, wxColour_Blue)(void* _obj)
 	
 EWXWEXPORT(int, wxColour_GetPixel)(void* _obj)
 {
-	return ((wxColour*)_obj)->GetPixel();
+  return (int)((wxColour*)_obj)->GetPixel();
 }
 
 EWXWEXPORT(void*, wxColour_Copy)(void* _obj)
